Renamed ui props interfaces and moved the fields for NewPatient to a separate file, new-patient-field
