+++ client/src/ehr_components/ui/textarea.tsx
@@ -1,13 +1,15 @@
-export interface Prop {
-  value: string
-  name: string
-  placeholder?: string
-  onChange: (name: string, event: React.ChangeEvent<HTMLTextAreaElement>) => void
-  className?: string
+export interface TextAreaProps {
+  value: string;
+  name: string;
+  placeholder?: string;
+  onChange: (
+    name: string,
+    event: React.ChangeEvent<HTMLTextAreaElement>
+  ) => void;
+  className?: string;
 }
 
-const TextArea = (props: Prop) => {
-
+const TextArea = (props: TextAreaProps) => {
   let { value, name, onChange, placeholder, className } = props;
 
   function _onChange(event: any) {
@@ -15,7 +17,7 @@ const TextArea = (props: Prop) => {
   }
 
   return (
-    <textarea 
+    <textarea
       className={`form-control ${className}`}
       id={name}
       value={value}
@@ -23,7 +25,7 @@ const TextArea = (props: Prop) => {
       onChange={_onChange}
       onInput={_onChange}
     />
-  )
-}
+  );
+};
 
 export default TextArea;
